adapting core waveform test
dev-buildbots on 64bit failed: http://tests.obspy.org/?id=1243#0

arange(10) returns an int64 on 64bit platforms and
mseed does not support int64 and will raise an exception. This is a wanted and
desired behaviour which will automatically adjusts the compression type based
on the given data (e.g. int32, float32, float64).

In contrast other formats support only one data type and we can savely
"require" to that type for these formats.

+++ obspy.core/trunk/obspy/core/tests/test_waveform_plugins.py
@@ -170,7 +170,8 @@ class WaveformPluginsTestCase(unittest.TestCase):
         formats_write = set(_getPlugins('obspy.plugin.waveform', 'writeFormat'))
         formats_read = set(_getPlugins('obspy.plugin.waveform', 'readFormat'))
         formats = set.intersection(formats_write, formats_read)
-        data = np.arange(10)
+        # mseed will raise exception for int64 data, thus use int32 only
+        data = np.arange(10, dtype='int32')
         # make array non-contiguous
         data = data[::2]
         tr = Trace(data=data)
